chore: fork upstream, strip down to just the Rust bits
This commit is a hard fork of
https://github.com/zxch3n/tidy/commit/54382fae3a9e85ac8329fa89d5a83632f20c2cde

We do the following:

1. Strip out all but the Rust+Wasm bits of upstream.

2. Rename the Rust packages `tidyt` (from `tidy-tree`) and
`tidyt-wasm` (from `wasm`).

3. Bump the version to `1.0.0`.

4. Add Nix packaging.

5. Port the `tidyt` package benchmarks to Criterion.

Other than the new package names and the Criterion port, the code in
this commit is exactly the same as in the upstream commit given above.
There are no functional changes.

Note that some of the package checks are not currently working, and
will require some additional work.

Signed-off-by: Drew Hess <src@drewhess.com>